requirement.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.7.0
appdirs==1.4.4
APScheduler==3.10.1
astroid==2.15.3
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
autopep8==2.0.2
backcall==0.2.0
behave==1.2.6
blinker==1.6.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==7.1.2
crcmod==1.7
cryptography==40.0.2
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
dicttoxml==1.7.16
dill==0.3.6
dnspython==2.3.0
docstring-to-markdown==0.12
ecdsa==0.13.3
enum34==1.1.10
executing==1.2.0
fastapi==0.97.0
flake8==6.0.0
Flask==2.3.2
Flask-Cors==3.0.10
frozenlist==1.3.3
future==0.18.3
gevent==22.10.2
greenlet==2.0.2
grpcio==1.54.0
grpcio-tools==1.54.0
h11==0.14.0
httpie==3.2.1
idna==3.4
importlib-resources==5.12.0
intelhex==2.3.0
ipaddress==1.0.23
ipython==8.12.0
isort==5.12.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jwcrypto==1.4.2
lazy-object-proxy==1.9.0
libusb1==1.10.1
limits==3.4.0
lxml==4.9.2
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
mccabe==0.7.0
mdurl==0.1.2
motor==3.2.0
msgpack==1.0.5
multidict==6.0.4
neovim==0.3.1
nrfutil==5.2.0
openai==0.27.8
packaging==23.1
parse==1.19.1
parse-type==0.6.1
parso==0.8.3
pc-ble-driver-py==0.11.4
pexpect==4.8.0
piccata==1.0.1
pickleshare==0.7.5
platformdirs==3.2.0
pluggy==1.0.0
prompt-toolkit==3.0.38
protobuf==4.23.4
ptpython==3.0.23
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.10.0
pycparser==2.21
pycryptodomex==3.9.4
pydantic==1.10.9
pydocstyle==6.3.0
pyflakes==3.0.1
Pygments==2.15.1
PyJWT==1.7.1
pylint==2.17.2
pymongo==4.4.0
pynvim==0.4.3
pyOpenSSL==23.1.1
pyserial==3.5
PySocks==1.7.1
pyspinel==1.0.0a3
python-dateutil==2.8.2
python-lsp-jsonrpc==1.0.0
python-lsp-server==1.7.2
pytoolconfig==1.2.5
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==4.2b4
redis==4.5.5
requests==2.30.0
requests-toolbelt==0.10.1
rich==13.3.4
rope==1.7.0
six==1.16.0
sniffio==1.3.0
snowballstemmer==2.2.0
SQLAlchemy==2.0.13
stack-data==0.6.2
starlette==0.27.0
toml==0.10.2
tomlkit==0.11.7
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
ujson==5.7.0
unicorn==2.0.1.post1
urllib3==1.26.15
uvicorn==0.22.0
wcwidth==0.2.6
websockets==11.0.3
Werkzeug==2.3.4
whatthepatch==1.0.4
wrapt==1.15.0
xmltodict==0.13.0
yapf==0.32.0
yarl==1.9.2
zope.event==4.6
zope.interface==6.0
